Shifts.

+++ src/main/shifts/shifts_controller.py
@@ -1,12 +1,10 @@
-from datetime import datetime
-
 from fastapi import APIRouter, Depends
-from sqlalchemy import extract
+from fastapi import HTTPException
+from sqlalchemy import extract, func
 from sqlalchemy.orm import Session
 
 from src.main.auth.auth_controller import get_current_user
 from src.main.commons.db_configuration import get_db
-from src.main.groups.models import Group, UserGroupRole
 from src.main.shifts.models import Shift
 from src.main.shifts.schemes import CreateShift
 from src.main.users.models import User
@@ -28,6 +26,15 @@ async def get_shifts(month: int, year: int, db: Session = Depends(get_db), user:
 
 @router.post("")
 async def create_shift(request: CreateShift, db: Session = Depends(get_db), user: User = Depends(get_current_user)):
+    shift_date = request.start_time.date()
+
+    existing_shift = db.query(Shift).filter(
+        Shift.user_id == user.id,
+        func.date(Shift.start_time) == shift_date
+    ).first()
+    if existing_shift:
+        raise HTTPException(status_code=400, detail="Ya tienes un turno para este día")
+
     shift: Shift = Shift(
         user_id=user.id,
         start_time=request.start_time,
@@ -44,6 +51,9 @@ async def create_shift(request: CreateShift, db: Session = Depends(get_db), user
 @router.delete("/{shift_id}")
 async def delete_group(shift_id: str, db: Session = Depends(get_db), user: User = Depends(get_current_user)):
     shift: Shift | None = db.query(Shift).filter(Shift.id == shift_id, Shift.user_id == user.id).first()
+    if shift is None:
+        return
+
     db.delete(shift)
     db.commit()
